implimented with DTO, bug fixes
