+++ app/src/main/java/com/nicos/carousel_compose/compose/pokemon_list_screen/PokemonListState.kt
@@ -1,10 +1,10 @@
-package com.nicos.pokedex_compose.compose.pokemon_list_screen
+package com.nicos.carousel_compose.compose.pokemon_list_screen
 
 import com.nicos.carousel_compose.data.room_database.entities.PokemonEntity
 
 data class PokemonListState(
     val pokemonMutableList: MutableList<PokemonEntity>? = null,
     var nextPage: String? = null,
-    val isLoading: Boolean = false,
+    val isLoading: Boolean = true,
     val error: String? = null,
 )
