+++ app/src/main/java/com/nicos/carousel_compose/compose/pokemon_list_screen/PokemonListViewModel.kt
@@ -4,7 +4,6 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.nicos.carousel_compose.domain.repositories.PokemonListRepository
 import com.nicos.carousel_compose.utils.generic_classes.Resource
-import com.nicos.pokedex_compose.compose.pokemon_list_screen.PokemonListState
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -26,47 +25,53 @@ class PokemonListViewModel @Inject constructor(
     }
 
     fun requestToFetchPokemon(url: String? = null) = viewModelScope.launch(Dispatchers.IO) {
-        _pokemonListState.value = _pokemonListState.value.copy(isLoading = true)
         pokemonListRepository.fetchPokemonList(url = url).let { resource ->
             when (resource) {
                 is Resource.Success -> {
-                    _pokemonListState.value =
-                        _pokemonListState.value.copy(
-                            isLoading = false,
-                            pokemonMutableList = resource.data,
-                                    nextPage = resource.nextUrl
-                        )
+                    viewModelScope.launch(Dispatchers.Main) {
+                        _pokemonListState.value =
+                            _pokemonListState.value.copy(
+                                isLoading = false,
+                                pokemonMutableList = resource.data,
+                                nextPage = resource.nextUrl
+                            )
+                    }
                 }
 
                 is Resource.Error -> {
-                    _pokemonListState.value =
-                        _pokemonListState.value.copy(
-                            isLoading = false,
-                            error = resource.message
-                        )
+                    viewModelScope.launch(Dispatchers.Main) {
+                        _pokemonListState.value =
+                            _pokemonListState.value.copy(
+                                isLoading = false,
+                                error = resource.message
+                            )
+                    }
                 }
             }
         }
     }
 
     private fun offline() = viewModelScope.launch(Dispatchers.IO) {
-        _pokemonListState.value = _pokemonListState.value.copy(isLoading = true)
         pokemonListRepository.offline().let { resource ->
             when (resource) {
                 is Resource.Success -> {
-                    _pokemonListState.value =
-                        _pokemonListState.value.copy(
-                            isLoading = false,
-                            pokemonMutableList = resource.data
-                        )
+                    viewModelScope.launch(Dispatchers.Main) {
+                        _pokemonListState.value =
+                            _pokemonListState.value.copy(
+                                isLoading = false,
+                                pokemonMutableList = resource.data
+                            )
+                    }
                 }
 
                 is Resource.Error -> {
-                    _pokemonListState.value =
-                        _pokemonListState.value.copy(
-                            isLoading = false,
-                            error = resource.message
-                        )
+                    viewModelScope.launch(Dispatchers.Main) {
+                        _pokemonListState.value =
+                            _pokemonListState.value.copy(
+                                isLoading = false,
+                                error = resource.message
+                            )
+                    }
                 }
             }
         }
