+++ app/src/main/java/com/nicos/carousel_compose/domain/dto/PokemonDto.kt
@@ -1,0 +1,8 @@
+package com.nicos.carousel_compose.domain.dto
+
+data class PokemonDto(
+    val name: String,
+    val url: String?,
+    var imageUrl: String?,
+    var order: Int?,
+)
