+++ app/src/main/java/com/nicos/carousel_compose/domain/pokemon_response_model/PokemonResponse.kt
@@ -1,10 +1,11 @@
-package com.nicos.carousel_compose.data.models.pokemon_response_model
+package com.nicos.carousel_compose.domain.pokemon_response_model
 
 import com.google.gson.annotations.SerializedName
 import com.nicos.carousel_compose.data.room_database.entities.PokemonEntity
+import com.nicos.carousel_compose.domain.dto.PokemonDto
 
 data class PokemonResponse(
     @SerializedName("next") val nextUrl: String?,
-    @SerializedName("results") val results: MutableList<PokemonEntity>
+    @SerializedName("results") val results: MutableList<PokemonDto>
 ) {
 }
