+++ app/src/main/java/com/nicos/carousel_compose/domain/repositories/PokemonListRepository.kt
@@ -1,6 +1,8 @@
 package com.nicos.carousel_compose.domain.repositories
 
+import androidx.core.text.isDigitsOnly
 import com.nicos.carousel_compose.data.room_database.entities.PokemonEntity
+import com.nicos.carousel_compose.data.room_database.entities.toPokemonEntity
 import com.nicos.carousel_compose.data.room_database.init_database.MyRoomDatabase
 import com.nicos.carousel_compose.domain.remote.PokemonService
 import com.nicos.carousel_compose.utils.generic_classes.HandlingError
@@ -14,13 +16,20 @@ class PokemonListRepository @Inject constructor(
     private val handlingError: HandlingError,
 ) {
 
+    companion object {
+        private const val BASE_IMAGE_URL =
+            "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/other/home/"
+        private const val PNG_FORMAT = ".png"
+    }
+
     suspend fun fetchPokemonList(url: String?): Resource<MutableList<PokemonEntity>> {
         return try {
             val pokemonService =
                 if (url == null) pokemonService.getPokemon() else pokemonService.getPokemon(url)
             val nextUrl = pokemonService.nextUrl
             val pokemonEntityList = pokemonService.results
-            savePokemon(pokemonEntityList = pokemonEntityList)
+            savePokemon(pokemonEntityList = pokemonEntityList.map { it.toPokemonEntity() }
+                .toMutableList())
 
             Resource.Success(data = myRoomDatabase.pokemonDao().getAllPokemon(), nextUrl = nextUrl)
         } catch (e: Exception) {
@@ -28,11 +37,23 @@ class PokemonListRepository @Inject constructor(
         }
     }
 
-    private suspend fun savePokemon(pokemonEntityList: MutableList<PokemonEntity>) =
-        PokemonEntity.savePokemonList(
-            pokemonEntityList = pokemonEntityList,
-            myRoomDatabase = myRoomDatabase
-        ).collect()
+    private suspend fun savePokemon(pokemonEntityList: MutableList<PokemonEntity>) {
+        pokemonEntityList.forEach {
+            buildPokemonImageUrl(it)
+            if (it.imageUrl != null) {
+                myRoomDatabase.pokemonDao().insertOrReplaceObject(it)
+            }
+        }
+    }
+
+    private fun buildPokemonImageUrl(pokemonEntity: PokemonEntity) {
+        val pokemonIdAsString =
+            pokemonEntity.url?.substringBeforeLast("/")?.substringAfterLast("/")
+        if (pokemonIdAsString != null && pokemonIdAsString.isDigitsOnly()) {
+            pokemonEntity.order = pokemonIdAsString.toInt()
+            pokemonEntity.imageUrl = "$BASE_IMAGE_URL$pokemonIdAsString$PNG_FORMAT"
+        }
+    }
 
     suspend fun offline(): Resource<MutableList<PokemonEntity>> {
         return try {
